# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "AppForMovies.API",
        "description": "This API provides services for renting and purchasing movies",
        "contact": {
            "name": "Software Engineering II Team",
            "email": "isii@on.uclm.es"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/Herramienta/Para-Oferta": {
            "get": {
                "tags": [
                    "Herramienta"
                ],
                "operationId": "GetHerramientaParaOferta",
                "parameters": [
                    {
                        "name": "fabricante",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaOfertarDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Herramienta/Para-Comprar": {
            "get": {
                "tags": [
                    "Herramienta"
                ],
                "operationId": "GetHerramientaParaComprar",
                "parameters": [
                    {
                        "name": "material",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientasParaComprarDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Herramienta/Para-Alquilar": {
            "get": {
                "tags": [
                    "Herramienta"
                ],
                "operationId": "GetHerramientasParaAlquilar",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "material",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaAlquilarDTO"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaAlquilarDTO"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaAlquilarDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Herramienta/Para-Reparar": {
            "get": {
                "tags": [
                    "Herramienta"
                ],
                "operationId": "GetHerramientasParaReparar",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tiempoReparacion",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaRepararDTO"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaRepararDTO"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HerramientaRepararDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HerramientaAlquilarDTO": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "nullable": true
                    },
                    "material": {
                        "type": "string",
                        "nullable": true
                    },
                    "fabricante": {
                        "type": "string",
                        "nullable": true
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
            "HerramientaRepararDTO": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "nullable": true
                    },
                    "material": {
                        "type": "string",
                        "nullable": true
                    },
                    "nombreFabricante": {
                        "type": "string",
                        "nullable": true
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "tiempoReparacion": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "additionalProperties": false
            },
            "HerramientasParaComprarDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nombre": {
                        "type": "string",
                        "nullable": true
                    },
                    "material": {
                        "type": "string",
                        "nullable": true
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    },
                    "fabricante": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "HerramientasParaOfertarDTO": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "nullable": true
                    },
                    "material": {
                        "type": "string",
                        "nullable": true
                    },
                    "fabricante": {
                        "type": "string",
                        "nullable": true
                    },
                    "precio": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}